+++ js/src/components/paid-ads/budget-section/index.js
@@ -9,7 +9,7 @@ import { __ } from '@wordpress/i18n';
 import Section from '.~/wcdl/section';
 import AppInputControl from '.~/components/app-input-control';
 import useStoreCurrency from '.~/hooks/useStoreCurrency';
-import getMonthlyMaxEstimated from '../getMonthlyMaxEstimated';
+import getMonthlyMaxEstimated from '../../../setup-ads/ads-stepper/create-campaign/getMonthlyMaxEstimated';
 import './index.scss';
 import FreeAdCredit from './free-ad-credit';
 import BudgetRecommendation from './budget-recommendation';
