+++ js/src/pages/edit-paid-ads-campaign/edit-paid-ads-campaign-form-content.js
@@ -8,6 +8,7 @@ import { __ } from '@wordpress/i18n';
  */
 import AudienceSection from '.~/components/paid-ads/audience-section';
 import FaqsSection from '.~/components/paid-ads/faqs-section';
+import BudgetSection from '.~/components/paid-ads/budget-section';
 
 const EditPaidAdsCampaignFormContent = ( props ) => {
 	const { formProps } = props;
@@ -22,6 +23,7 @@ const EditPaidAdsCampaignFormContent = ( props ) => {
 				) }
 				formProps={ formProps }
 			/>
+			<BudgetSection formProps={ formProps } />
 			<FaqsSection />
 		</>
 	);
