+++ js/src/setup-ads/ads-stepper/create-campaign/form-content.js
@@ -8,8 +8,8 @@ import { __ } from '@wordpress/i18n';
  */
 import StepContentFooter from '.~/components/stepper/step-content-footer';
 import AudienceSection from '.~/components/paid-ads/audience-section';
-import BudgetSection from './budget-section';
-import FaqsSection from '../../../components/paid-ads/faqs-section';
+import BudgetSection from '.~/components/paid-ads/budget-section';
+import FaqsSection from '.~/components/paid-ads/faqs-section';
 
 const FormContent = ( props ) => {
 	const { formProps, submitButton } = props;
